[watchOS] refactor: 폴더 구조화 및 Mondee 구조 분리 (#3)
